Fix bug in `totalChildren` counts
Previously, `totalChildren` was only getting set to a
non-zero value for root threads (it was 0 for all non-root
threads, which isn't helpful). More pressingly, it
didn't recalculate `totalChildren` after children
were removed due to filtering. This is fixed here.

`totalChildren` is no longer set on non-root threads. It
does not appear to have been used, anyway (and, as it
was always `0` before, wouldn't have been useful).